refactor(workers): create workers package
Move timeout management from threading/ to workers/ package and add

Changes:
- Move threading/timeout.py → workers/timeout.py
- Update __main__.py import to use new location
- Move tests/threading/ → tests/workers/
